+++ lib/ch/query.ex
@@ -232,6 +232,7 @@ defimpl DBConnection.Query, for: Ch.Query do
   defp encode_param(%Decimal{} = d), do: Decimal.to_string(d, :normal)
   defp encode_param(%Date{} = date), do: Date.to_iso8601(date)
   defp encode_param(%NaiveDateTime{} = naive), do: NaiveDateTime.to_iso8601(naive)
+  defp encode_param(%Time{} = time), do: Time.to_iso8601(time)
 
   defp encode_param(%DateTime{microsecond: microsecond} = dt) do
     dt = DateTime.shift_zone!(dt, "Etc/UTC")