+++ lib/ch/row_binary.ex
@@ -98,6 +98,7 @@ defmodule Ch.RowBinary do
               :date,
               :datetime,
               :date32,
+              :time,
               :ipv4,
               :ipv6,
               :point,
@@ -155,6 +156,8 @@ defmodule Ch.RowBinary do
   defp encoding_type({:datetime64, _, tz}) do
     raise ArgumentError, "can't encode DateTime64 with non-UTC timezone: #{inspect(tz)}"
   end
+
+  defp encoding_type({:time64 = t, p}), do: {t, time_unit(p)}
 
   defp encoding_type({e, mappings}) when e in [:enum8, :enum16] do
     {e, Map.new(mappings)}
@@ -321,11 +324,11 @@ defmodule Ch.RowBinary do
     <<DateTime.diff(datetime, @epoch_utc_datetime, time_unit)::64-little-signed>>
   end
 
-  def encode({:datetime64, _precision}, %DateTime{} = datetime) do
+  def encode({:datetime64, _time_unit}, %DateTime{} = datetime) do
     raise ArgumentError, "non-UTC timezones are not supported for encoding: #{datetime}"
   end
 
-  def encode({:datetime64, _precision}, nil), do: <<0::64>>
+  def encode({:datetime64, _time_unit}, nil), do: <<0::64>>
 
   def encode(:date, %Date{} = date) do
     <<Date.diff(date, @epoch_date)::16-little>>
@@ -338,6 +341,29 @@ defmodule Ch.RowBinary do
   end
 
   def encode(:date32, nil), do: <<0::32>>
+
+  def encode(:time, %Time{} = time) do
+    {s, _micros} = Time.to_seconds_after_midnight(time)
+    <<s::32-little-signed>>
+  end
+
+  def encode(:time, nil), do: <<0::32>>
+
+  def encode({:time64, time_unit}, %Time{} = time) do
+    {s, micros} = Time.to_seconds_after_midnight(time)
+
+    micros_as_ticks =
+      cond do
+        time_unit < 1_000_000 -> div(micros, time_unit)
+        time_unit == 1_000_000 -> micros
+        true -> micros * div(time_unit, 1_000_000)
+      end
+
+    ticks = s * time_unit + micros_as_ticks
+    <<ticks::64-little-signed>>
+  end
+
+  def encode({:time64, _time_unit}, nil), do: <<0::64>>
 
   def encode(:uuid, <<u1::64, u2::64>>), do: <<u1::64-little, u2::64-little>>
 
@@ -551,6 +577,8 @@ defmodule Ch.RowBinary do
               :uuid,
               :date,
               :date32,
+              :time,
+              :time64,
               :ipv4,
               :ipv6,
               :point,
@@ -593,6 +621,8 @@ defmodule Ch.RowBinary do
 
   defp decoding_type({:datetime64 = t, p}), do: {t, time_unit(p), _tz = nil}
   defp decoding_type({:datetime64 = t, p, tz}), do: {t, time_unit(p), tz}
+
+  defp decoding_type({:time64 = t, p}), do: {t, time_unit(p)}
 
   defp decoding_type({e, mappings}) when e in [:enum8, :enum16] do
     {e, Map.new(mappings, fn {k, v} -> {v, k} end)}
@@ -877,6 +907,42 @@ defmodule Ch.RowBinary do
       :date32 ->
         <<d::32-little-signed, bin::bytes>> = bin
         decode_rows(types_rest, bin, [Date.add(@epoch_date, d) | row], rows, types)
+
+      :time ->
+        <<s::32-little-signed, bin::bytes>> = bin
+
+        t =
+          if s >= 0 and s < 86400 do
+            Time.from_seconds_after_midnight(s)
+          else
+            # since ClickHouse supports Time values of [-999:59:59, 999:59:59]
+            # and Elixir's Time supports values of [00:00:00, 23:59:59]
+            # we raise an error when ClickHouse's Time value is out of Elixir's Time range
+            raise ArgumentError,
+                  "ClickHouse Time value #{s} (seconds) is out of Elixir's Time range (00:00:00 - 23:59:59)"
+
+            # TODO: we could potentially decode ClickHouse's Time values as Elixir's Duration when it's out of Elixir's Time range
+          end
+
+        decode_rows(types_rest, bin, [t | row], rows, types)
+
+      {:time64, time_unit} ->
+        <<ticks::64-little-signed, bin::bytes>> = bin
+
+        t =
+          if ticks >= 0 and ticks < 86400 * time_unit do
+            ticks |> DateTime.from_unix!(time_unit) |> DateTime.to_time()
+          else
+            # since ClickHouse supports Time64 values of [-999:59:59.999999999, 999:59:59.999999999]
+            # and Elixir's Time supports values of [00:00:00.000000, 23:59:59.999999]
+            # we raise an error when ClickHouse's Time64 value is out of Elixir's Time range
+            raise ArgumentError,
+                  "ClickHouse Time value #{ticks / time_unit} (seconds) is out of Elixir's Time range (00:00:00.000000 - 23:59:59.999999)"
+
+            # TODO: we could potentially decode ClickHouse's Time64 values as Elixir's Duration when it's out of Elixir's Time range
+          end
+
+        decode_rows(types_rest, bin, [t | row], rows, types)
 
       {:datetime, timezone} ->
         <<s::32-little, bin::bytes>> = bin