+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- add experimental [Time](https://clickhouse.com/docs/sql-reference/data-types/time) and [Time64](https://clickhouse.com/docs/sql-reference/data-types/time64) types support https://github.com/plausible/ch/pull/260
+
 ## 0.4.1 (2025-07-07)
 
 - fix column decoding when count exceeds 127 https://github.com/plausible/ch/pull/257